Interogările din /piese pleacă în două valuri, nu în cinci pe rând
Erau cinci `await` unul după altul, deci cinci drumuri dus-întors până la Supabase, în
serie. Doar interogarea de piese are nevoie de rezultatele celorlalte — traduce slug-urile
din adresă în id-uri, prin cats.find, models.find și brands.find — așa că restul pot pleca
deodată. Verificat înainte de a le muta, ca să nu rup dependența.

Măsurat local, A/B pe aceeași infrastructură, cinci rulări fiecare:
  /piese              secvențial 1,08 s -> paralel 0,78 s
  /piese?marca=skoda  secvențial 1,03 s -> paralel 0,71 s

Erorile din valul 1 NU sunt înghițite. `Promise.all` respinge la prima eroare și n-o
prindem: pagina cade vizibil, cu 500. Dacă am fi pus `?? []` pe fiecare rezultat, o
interogare căzută ar fi dat o listă goală de mărci sau de categorii, iar pagina ar fi
scris „nicio piesă găsită" ca și cum ar fi fost adevărat — același tipar ca la plafonul de
1.000 de rânduri și ca la RLS. Verificat rupând intenționat o interogare: HTTP 500, fără
„piese găsite" și fără „Nicio piesă" în răspuns.

`numar_piese_pe_model` trece și el prin citesteTot: are 538 de rânduri azi, dar creștea
odată cu modelele și n-avea nici paginare, nici tratare de eroare.

Migrarea 31 e aplicată; fișierul ei poartă acum și motivul pentru care nu are `::int`:
`create or replace view` nu poate schimba tipul unei coloane, iar `count(*)` vechi era
bigint. Capcana nu se vede testând definiția sub alt nume.

+++ supabase/categorii-numar-rapid.sql
@@ -38,11 +38,22 @@
 -- `security_invoker = true` se păstrează, din același motiv ca la migrarea 14:
 -- view-ul aplică politicile RLS ale celui care îl citește. Se redeclară pentru că
 -- `create or replace view` NU păstrează opțiunile view-ului vechi.
+--
+-- ⚠ DE CE NU E `count(distinct pid)::int`
+-- Prima variantă avea cast la `int` și a fost respinsă:
+--
+--   ERROR: 42P16: cannot change data type of view column „nr_piese"
+--                 from bigint to integer
+--
+-- `create or replace view` poate schimba CORPUL unui view, dar nu tipul unei
+-- coloane. Vechiul `count(*)` întorcea `bigint`, deci noul trebuie tot `bigint`.
+-- Capcana nu se vede testând definiția sub alt nume: acolo e un view NOU, iar
+-- restricția nu se aplică. Se vede doar la înlocuirea propriu-zisă.
 -- ============================================================
 
 create or replace view public.categorii_cu_numar as
   with numarate as (
-    select cid, count(distinct pid)::int as nr
+    select cid, count(distinct pid) as nr
       from (
         select id as pid, categorie_id    as cid from products
          where publicat and stoc > 0 and categorie_id is not null
